Remove ZoomButtonsController view immediately when !visible
When an acitvity with ZoomButtonsController is destorying all
its views are removed immediately except for the
ZoomButtonController view which is deferred. This causing the
system to think the window was leaked at the end of the current
clean-up pass since the view is still around. We now remove
the view immediately when setVisible(false) is called which is
the current signal that the window is detaching or the parent
view is now hidden.

Bug: 18280446
Change-Id: If566b402c2914a72539e28e4f6d768c6199f2d2f